test: PieceTest를 각 Piece 별로 분리
